permissions

+++ src/Controller/MaladieController.php
@@ -19,10 +19,6 @@ class MaladieController extends AbstractController
     public function index(EntityManagerInterface $entityManager): Response
     {
         $usr = $this->getUser();
-        if(!isset($usr->getRoles()['ROLE_ADMIN'])) {
-            $this->user = $usr;
-            return $this->redirectToRoute('app_club_show', ["id" => $usr->getClubid()->getId()], Response::HTTP_SEE_OTHER);
-        }
         if(isset($usr->getRoles()['ROLE_MASTER'])){
             $maladies = $entityManager
                 ->getRepository(Maladie::class)
@@ -47,6 +43,10 @@ class MaladieController extends AbstractController
     public function new(Request $request,EntityManagerInterface $entityManager): Response
     {
         $usr = $this->getUser();
+        if(!isset($usr->getRoles()['ROLE_ADMIN']) && !isset($usr->getRoles()['ROLE_DOCTOR'])) {
+            $this->user = $usr;
+            return $this->redirectToRoute('app_club_show', ["id" => $usr->getClubid()->getId()], Response::HTTP_SEE_OTHER);
+        }
         $maladie = new Maladie();
         $form = $this->createForm(MaladieType::class, $maladie);
         $form->handleRequest($request);
@@ -97,6 +97,11 @@ class MaladieController extends AbstractController
     public function edit(Request $request, Maladie $maladie, EntityManagerInterface $entityManager): Response
     {
         $usr = $this->getUser();
+
+        if(!isset($usr->getRoles()['ROLE_ADMIN']) && !isset($usr->getRoles()['ROLE_DOCTOR'])) {
+            $this->user = $usr;
+            return $this->redirectToRoute('app_club_show', ["id" => $usr->getClubid()->getId()], Response::HTTP_SEE_OTHER);
+        }
         $form = $this->createForm(MaladieType::class, $maladie);
         $form->handleRequest($request);
         $sections = $entityManager
@@ -121,6 +126,10 @@ class MaladieController extends AbstractController
     public function delete(Request $request, Maladie $maladie, EntityManagerInterface $entityManager): Response
     {
         $usr = $this->getUser();
+        if(!isset($usr->getRoles()['ROLE_ADMIN']) && !isset($usr->getRoles()['ROLE_DOCTOR'])) {
+            $this->user = $usr;
+            return $this->redirectToRoute('app_club_show', ["id" => $usr->getClubid()->getId()], Response::HTTP_SEE_OTHER);
+        }
         if ($this->isCsrfTokenValid('delete'.$maladie->getId(), $request->request->get('_token'))) {
             $entityManager->remove($maladie);
             $entityManager->flush();
